IDEMPIERE-5567 Support of UUID as Key (FHCA-4195) - Translation (#2119)
* IDEMPIERE-5567 Support of UUID as Key (FHCA-4195)

- Manage translation for UUID based tables
- Enable 2Pack Import UUID based tables with translation
- Create ad_package_imp_detail.Record_UU
- Create test table TestUU_Trl

* - Fix Synchronize Doc Translation

+++ org.adempiere.pipo.handlers/src/org/adempiere/pipo2/handler/ViewColumnElementHandler.java
@@ -54,8 +54,8 @@ public class ViewColumnElementHandler extends AbstractElementHandler {
 			MViewColumn mViewColumn = findPO(ctx, element);
 			if (mViewColumn == null) {
 				int parentId = 0;
-				if (getParentId(element, MViewComponent.Table_Name) > 0) {
-					parentId = getParentId(element, MViewComponent.Table_Name);
+				if ((Integer)getParentId(element, MViewComponent.Table_Name) > 0) {
+					parentId = (Integer)getParentId(element, MViewComponent.Table_Name);
 				} else {
 					Element pfElement = element.properties.get(MViewColumn.COLUMNNAME_AD_ViewComponent_ID);
 					parentId = ReferenceUtils.resolveReferenceAsInt(ctx.ctx, pfElement, getTrxName(ctx));
